src/app/requirements.txt
dash==2.18.2
supabase==2.11.0
python-dotenv==1.0.1
gunicorn==23.0.0
